Move jobStartedCompleteFromPopupDialog earlier in transform flow.
